requirements.txt
pyserial==3.4
influxdb==5.3.0
